projets/P11-rag-system-mistral/Scripts:Tests/requirements.txt
langchain
langchain-community
langchain-mistralai
mistralai
faiss-cpu
pandas
numpy
python-dotenv
pytest
jupyter
requests